\name{bcfa2s-package}
\alias{bcfa2s-package}
\alias{bcfa2s}
\docType{package}
\title{
\packageTitle{bcfa2s}
}
\description{
\packageDescription{bcfa2s}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bcfa2s}
\packageIndices{bcfa2s}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bcfa2s}

Maintainer: \packageMaintainer{bcfa2s}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
