% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_spa.R
\name{estim_spa}
\alias{estim_spa}
\title{Estimate the structural parameters of the model}
\usage{
estim_spa(
  y,
  L,
  lam_ad,
  lambda_mean,
  psi_mean,
  intercepts = TRUE,
  n.iter = 1000,
  n.thin = 1,
  n.burnin = floor(n.iter/2),
  n.chains = 3
)
}
\arguments{
\item{n.iter}{total number of iterations}

\item{n.thin}{thinning factor}

\item{n.burnin}{the number of iterations to burn}
}
\value{
an \code{mcmc} object
}
\description{
Estimate the structural parameters of the model
}
