% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_thresholds.R
\name{find_thresholds}
\alias{find_thresholds}
\title{Estimate the thresholds from the empirical distributions}
\usage{
find_thresholds(x)
}
\arguments{
\item{x}{an ordinal vector}
}
\value{
a vector of thresholds
}
\description{
Estimate the thresholds from the empirical distributions
}
