% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mu.R
\name{simulate_mu}
\alias{simulate_mu}
\title{Create replicates of the dataset with continuous 'latent' variables consistent with the ordinal data}
\usage{
simulate_mu(data, mvs, thresholds, sigma, nrep = 20L)
}
\description{
Create replicates of the dataset with continuous 'latent' variables consistent with the ordinal data
}
